+++ Resources/Locale/en-US/_Mono/armor-plate.ftl
@@ -10,6 +10,10 @@ armor-plate-examinable-verb-message = Examine protection and durability characte
 armor-plate-attributes-examine = This armor plate:
 armor-plate-initial-durability = Is rated for [color=yellow]{ $durability }[/color] standard units of damage.
 
+armor-plate-gait-speed = speed
+armor-plate-gait-walk = walking speed
+armor-plate-gait-sprint = running speed
+
 armor-plate-speed-display =
     { $deltasign ->
         [-1] Increases your {$gait} by [color=yellow]{$speedPercent}%[/color].
